flytt PdlClientTest til pakke pdl. Flytt pdl test data til PdlTestData
